+++ src/imports.h
@@ -1,7 +1,13 @@
 #ifndef IMPORTS_H
 #define IMPORTS_H
 
-#include "so_util.h"
+#include <stddef.h>
+#include <stdint.h>
+
+typedef struct {
+  const char *symbol;
+  uintptr_t func;
+} DynLibFunction;
 
 extern DynLibFunction dynlib_functions[];
 extern const int dynlib_functions_count;